rst: create a subdirectory to hold rst files
Using the current directory to hold ReStructuredText files can lead
to some funny errors.

Create a new directory "rst" to contain all such files, and specify
it as the input directory for the "sphinx-build" command in the
Makefile.  Move "conf.py" into that directory, and move the top
level "index.rst" and "participants.rst" (which it references) there
as well.

Signed-off-by: Alex Elder <elder@umn.edu>