docs: Manage each section of the docs site recursively
Rather than putting all articles in the main index.rst
and cluttering the site, we create sub-indexes in each
section. This not only cleans up the site visually, but
also has the added benefit of making it much easier to
maintain your own section without having to muck with the
main index.rst file.

Signed-off-by: Shaun Loo <loo00013@umn.edu>