docs: Recursively generate subsections
Rather than dumping everything into the main index.rst, we have a
subindex for every section (folder) in the docs repo, and have
the main index.rst point to these subsections. This not only
makes it easier for subgroups to maintain their own section of
the website, it also generates nested entries on the site itself,
making it look a lot cleaner.

Signed-off-by: Shaun Loo <loo00013@umn.edu>
Signed-off-by: Jake Freeland <freel238@umn.edu>